test(pipeline): move pipeline test modules out of router_factory.rs
router_factory.rs carried two test modules exercising pipeline
behavior: the plugin-instantiation matrices driven through
PipelineFactory::create_pipeline and the build_subgraph_services
tests. They move to pipeline/tests/plugin_instantiation.rs and
pipeline/tests/subgraph_services.rs, unchanged in body, leaving
router_factory.rs with only the traits and the factory. The
cfg(test) inject_schema_id re-export goes with them, since the moved
tests were its only external caller.
