+++ apollo-router/src/pipeline/mod.rs
@@ -32,10 +32,6 @@ pub(crate) use self::acquire::connect_query_plan_redis;
 pub(crate) use self::acquire::create_query_planner_service;
 pub(crate) use self::acquire::parse_http_client_material;
 pub(crate) use self::plugins::create_plugins;
-// Only tests outside this module call `inject_schema_id`; within the pipeline it is an
-// implementation detail of the acquire phase.
-#[cfg(test)]
-pub(crate) use self::plugins::inject_schema_id;
 pub(crate) use self::stages::SupergraphPipeline;
 pub(crate) use self::stages::build_apq_expander;
 pub(crate) use self::stages::build_http_services;
