+++ apollo-router/src/pipeline/tests.rs
@@ -12,6 +12,9 @@ use crate::services::SupergraphRequest;
 use crate::spec::Schema;
 use crate::uplink::license_enforcement::LicenseState;
 
+mod plugin_instantiation;
+mod subgraph_services;
+
 /// Subgraph names in `testdata/supergraph.graphql`, sorted.
 const FIXTURE_SUBGRAPHS: [&str; 4] = ["accounts", "inventory", "products", "reviews"];
